requirements.txt
streamlit
pandas
openpyxl
requests
google-genai
python-dotenv
pdfplumber
gspread
oauth2client
